test: add tests for hasSanityPackageInImportMap
